Use .balign in assembler.
On some platforms, .align is synonymous with .balign, and so

.align 16

is the same as

.balign 16

But on some platforms, gas likes to have fun with us, and it treats it
as .p2align instead, in which case that statement means

.balign 1<<16

Which is obviously always wrong.

Some people wonder why other people don't like the gnu toolchain.

Signed-off-by: Peter Jones <pjones@redhat.com>