Make efivar's vars.c interface choose between 32 and 64 bit at runtime.
We don't actually know what size the /kernel/ interface will use until
we're executed, because it's defined as "unsigned long", and we don't
know if a 32-bit build will be run on a 32-bit or 64-bit kernel.  So in
this case, we figure it out from uname.

This makes the assumption that all the 64-bit arches we might care about
(ia64, x86_64, and aarch64) have "64" in "uname -m" somewhere, and that
none of the 32-bit architectures do.

This should resolve the conflict between 487a9ea and a01d106, and
resolve issue #12 .

Signed-off-by: Peter Jones <pjones@redhat.com>