Try to actually discover the size of the vars struct by reading one.
Turns out kernel's commit e33655a38 makes /it/ try to decide which
struct to give based on if CONFIG_COMPAT is on and if the userland
process is using it.  So now we have no idea which thing we'll get.

Since the constant size is the problem in the first place, just try to
read one until we have all its data, and return based on the size.

If they all fail, we determine based on __SIZEOF_POINTER__ == 4 ,
but you're probably not going to get a lot of utility out of the
variables anyway.

Signed-off-by: Peter Jones <pjones@redhat.com>