Work around LocateDevicePath() not grokking PcieRoot() devices properly.
Basically: PNP0A03 always has to work, so everybody always uses it, so
nobody tests PNP0A08 working, so it doesn't.

So you have to lie to the machine about which hardware you want it to
find things on, because if you tell it the truth, it'll lie to you.

The truth is a lie.

Signed-off-by: Peter Jones <pjones@redhat.com>