libefiboot: figure out nvme device paths better.
Figure out that an nvme device is the nvme device and make the path to a
partition correctly.

Something is also wrong on the failure path there - we should fail with
an error if we legit can't figure out a device path and never get valid
sizes etc., not write a bogus HD() paths filled with zeros.

Signed-off-by: Peter Jones <pjones@redhat.com>