Make guids be in correct byte order both in text and ram...
This is all because the EFI spec defines:

typedef struct {
	UINT32 Data1;
	UINT16 Data2;
	UINT16 Data3;
	UINT8
	Data4[8];
} EFI_GUID;

	{0x8BE4DF61,0x93CA,0x11d2,\
		{0xAA,0x0D,0x00,0xE0,0x98,0x03,0x2B,0x8C}}

but prints:
8BE4DF61-93CA-11d2-AA0D-00E098032B8C

But we decided to be "smart" and did:

typedef struct {
	uint32_t        a;
	uint16_t        b;
	uint16_t        c;
	uint16_t        d;
	uint8_t         e[6];
} efi_guid_t;

Which means d is effectively in big endian while a, b, and c are little
endian.  This is awful.

We also have efi_guid_t in the public header, so things can take
sizeof() on it, which seems ill advised in hindsight.

Thankfully, we also have the EFI_GUID() macro, which was getting things
right, so if things memcmp()'d or efi_guid_cmp()'d on that, it'd get it
right, and efibootmgr is, in fact, using that (via EFI_GUID_GLOBAL)
instead.

So now, keep them all in memory as le32-le16-le16-be16-b-b-b-b-b-b at
all times.  The only programs that should break were ther ones that were
already broken, and they'll need a rebuild.

Also fix a bug introduced in c0a2b3db1088af888d98c2223a648bc6f6274eea
which introduced at "1" into some definitions of efi_guid_empty.

(yes, it really says "d2" not "D2")

Signed-off-by: Peter Jones <pjones@redhat.com>