creator/find_file(): fix an incorrect error that gets raised.
In find_file(), when we're trying to find the device that matches a
mountpoint, we incorrectly raise an error when we find an entry that's
got a longer name than the one we want.  This is incorrect - it's a
special case of what Lans Zhang was fixing in 8eac2d51, in fact.

This case should just reject *that* entry, not set an error and end the
iteration.

Signed-off-by: Peter Jones <pjones@redhat.com>