Replace popt usage with getopt_long in efivar.c.
efivar does not make use of popt in a way that really requires the
dependency.  Since this is a low level system utility, it's possible
some system builders want to reduce redundant libraries they need to
build and install.  popt is arguably redundant with getopt_long.  It's
possible to just drop long option handling entirely so it works with
plain old getopt.  The intent with this patch was to provide the same
command line options but just using what the C library provides rather
than popt.

The attributes variable in main() has also been changed to a uint32_t
size and sign consistency with its usage in edit_variable().

Signed-off-by: David Cantrell <david.l.cantrell@gmail.com>