NVDIMM: UUID<->GUID needs endian swizzling.
The NVDIMM Label in sysfs is a UUID rather than a GUID.  The big
difference is UUIDs are stored opposite-endian from GUIDs, so
efi_str_to_guid() is giving us back a GUID with the first 8 bytes in the
wrong order.  Fix it up after we read it from sysfs and parse it.

Signed-off-by: Peter Jones <pjones@redhat.com>