efivarfs / vars: usleep() before reading from efivarfs if euid != 0
There's a kernel rate limiter on efi variable reads now for
non-root users, and we'd rather just not hit it than have to dig out
from having hit it.  So this adds a 10ms sleep before each read call.

If you do have 50 variables, efibootmgr will do 100 reads, which would
trigger the rate limit.  In that case, this patch adds 1 second (plus
lossage due to calling, etc.), so it should stay just below the
triggering threshold.  That will definitely be /smoother/ than hitting
it, and almost certainly faster as well, because the extra calls will
re-enforce the limit.

Signed-off-by: Peter Jones <pjones@redhat.com>