If we can't parse part of the device link, skip it and set DEV_ABBREV_ONLY
If we can't parse some part of the device symlink, we can't write a full
device path, but we can write an abbreviated HD() or File() path.  So if
we've exausted all possibilities, skip to the next node, set
DEV_ABBREV_ONLY in the device's flags, and try parsing again.  Then when
creator.c checks if that flag conflicts, it'll throw an error if it
does.

Signed-off-by: Peter Jones <pjones@redhat.com>