Fix incorrect failure on absence of eui.
Do not fail when there's no eui on an nvme device. This is a bug
introduced in ff696a432b. nvme devices do not have to have eui.

The previous version was correct and there's has_eui to indicate whether
eui was successfully read.

The issue causes efi_generate_file_device_path_from_esp() to fail on all
the platforms which have ESP on an NVMe device.