Fix up efi_guid_cmp()'s alignment problem a different way
With the prior fix for efi_guid_cmp()'s alignment issue, abicheck shows:

1 function with some indirect sub-type change:

  [C]'function int efi_guid_cmp(void* const, void* const)' at <built-in>:34:1 has some indirect sub-type changes:
    parameter 1 of type 'void* const' changed:
      entity changed from 'void* const' to 'const efi_guid_t*'
      type size hasn't changed
    parameter 2 of type 'void* const' changed:
      entity changed from 'void* const' to 'const efi_guid_t*'
      type size hasn't changed

While this isn't a meaningful ABI difference in terms of linking, it is
definitely worse than having the type actually specified.

This patch changes the type back to the previous type, but also changes
the typedef to require a 1-byte alignment.  This will guarantee that all
new builds of efi_guid_cmp() and related code have code generated in a
way that's compatible with any alignment, thus alleviating the issue.

Signed-off-by: Peter Jones <pjones@redhat.com>