Try to deal with some signof(char) != signof(uint8_t) madness.
The 3-sign "char" type in C is so so painful sometimes.  This attempts
to use -funsigned-char to avoid type-casting between (uint8_t *) and
(char *) quite so much.  I'm not sure it'll pay off.  The fact that
"char" is something that's not ever actually the same type as "unsigned
char" or "signed char" is still the absolute worst thing.

Why do we need signed char at all again?

Signed-off-by: Peter Jones <pjones@redhat.com>