Fix a case clang-analyzer found where we may try to parse a NULL
I don't think this is something that can *actually* happen - it didn't
trigger before save_variable() was added, and the save_variable() path
that calls this calls validate_name() immediately prior to this call.
validate_name() calls exit() if it's NULL.  But that's weird as well,
because that's the same pattern all the other users of parse_name() use.

Anyway, this patch expands validate_name() and moves it into
parse_name() so we don't need to call it from everywhere when we're just
calling the two in a row anyway.

Signed-off-by: Peter Jones <pjones@redhat.com>