Work around autoconf existing in the world.
Long long ago, autoconf added a variable called to pass linker flags to
the compiler.  For some misguided reason, they prefixed it with LD (the
linker) instead of "CC" (the compiler).  Now people unquestioningly
complain if you use that name a more reasonable way, so we're stuck with
names that are counter-descriptive.  You just can't win, the well is too
full of poison.

Resolves github issue #142.

Signed-off-by: Peter Jones <pjones@redhat.com>