Add efi_time_t and time conversion and formatting utilities
This adds the hilariously defective efi_time_t type, as well as
conversion utilities to and from it, and utilities for formatted output
of it.

Optionally, for compatibility with older source trees, if your
application defines EFIVAR_NO_EFI_TIME_T to any value other than 0, it
will *not* do so, though that option may go away in some future efivar
version.  This allows you to turn off declaration of efi_time_t and
related functions in the case where it's declared someplace else, such
as in some local code or another library's headers.

Signed-off-by: Peter Jones <pjones@redhat.com>