hexdump: make it possible to display relative positions
When we're doing a hexdump, most of the time people don't care that much
about the memory address of the data in program memory, especially if
it's a hexdump of a file on disk similar data.

This adds hexudmpat(), which allows us to skew the position reported in
the dump by a constant factor, so that we can show positions relative to
the structure in memory rather than the address in ram itself.

Signed-off-by: Peter Jones <pjones@redhat.com>