efivar: move show_errors() to util.h
show_errors() will be used in utilities other than efivar, so it needs
to be in shared code.

This patch moves it to util.h, as well as changing it to use the
verbosity setting API that libefivar provides.  Right now it can't live
in a .c file unless that file /and error.c/ are linked into makeguids,
which currently results in dependency loops that break the build, so
it's still in util.h .

Signed-off-by: Peter Jones <pjones@redhat.com>