_vars_chmod_variable(): always expect the directory in "path"
Looks like the _vars_chmod_variable function has had this bug since it
was introduced in 2014. This function is called from 2 paths, the
vars_chmod_variable function, and vars_set_variable.

_vars_chmod_variable currently expects the filename of the data file to
be passed in, however, the 2 functions that call it pass in the parent
directory path.

We've applied this patch to our yocto rootfs and are now able to
successfully set permissions on efivar variables (using the old efivar
sysfs interface).