Makefile: Fix building thread-test with parallel jobs
The thread-test program does not declare a dependency on libefivar,
so it would fail to build if the library does not exist yet:

/usr/libexec/gcc/riscv64-gentoo-linux-gnu/ld: cannot find -lefivar
collect2: error: ld returned 1 exit status

Also drop the CFLAGS declaration that mixes options from different
platforms.  Since the program gets installed, stick with the target
system compiler and flags.

Signed-off-by: David Michael <fedora.dm0@gmail.com>