efivar: Fix guid sort order
For a long time, we've compared guids and their symbolic names with
memcmp().  That provides a stable sort order, and very little
(if anything) depends on the sorted order of a list of GUIDs for
anything other than searching.

Unfortunately, when you're actually looking at the sorted output of a
list of guids, this winds up being incredibly confusing, because the
first element of a guid is a little endian uint23_t, which means
memcmp() is evaluating the bytes in a different order than sorting a
text file full of guids would.

This patch switches to comparing the internal integral members of the
guid as unsigned integers, which is still slightly confusing when it
comes to the 4th one (it is compared opposite of how it appears, because
GUIDs are terrible), and to comparing the symbolic names with strncmp().

Signed-off-by: Peter Jones <pjones@redhat.com>