efivar: Fix incorrect error text
In the efivar utility, we're accidentally printing '-t' in the error
message for the '-A' argument.  This is clearly less than good.  We're
also printing the error code /twice/, presumably due to missing it in a
switch from fprintf() to err().

This fixes it to say '-A' and removes the extra output.

Signed-off-by: Peter Jones <pjones@redhat.com>