Minor fixes for clangd on some local headers
hexdump.h and util.h are both normally included from efivar.h, which
also pulls in all of the system headers that are normally needed for it.
Unfortunately, if you're editing the headers themselves, validation
tools don't know that's going on.

This change adds header includes for the things needed in those files.
This shouldn't change anything at all for this code when it's being used
as it always had been, because those headers' guard defines are already
there, but it makes it possible to run checkers on them individually.

Signed-off-by: Peter Jones <pjones@redhat.com>