Don't print newlines when printing structured dump output
At a couple of places, such as in our hexdump() functions, we use
isprint() to determine if something is reasonable to print, but
isprint('\n') returns nonzero, and we really don't want to print that
sort of thing there.

This adds a utility function safe_to_print(), which is an isprint()
wrapper that filters those out.

Signed-off-by: Peter Jones <pjones@redhat.com>