get_partition_info(): fix stack overflow
Coverity correctly noticed that sometimes we're plausibly assigning a
uint32_t to a spot on the stack that's only a uint8_t, which isn't
great.

This patch adds a new union type for the partition signature, and takes
out the broken type-casting that hides the error.  Unfortunately it
still has to leave cast in place on the efidp_make_hd() call, as that's
part of our ABI.

Signed-off-by: Peter Jones <pjones@redhat.com>