File device paths: don't print "File(" or ")"
ardb poked me and said:

<ardb> pjones: were you aware that EDK2 does not actually use File() for
       EFIDP_MEDIA_FILE device path nodes?
<ardb> e.g.,  parsing the following using the text-to-device-path
       protocol
<ardb> HD(1,GPT,400774eb-3297-4c12-906a-f913b4c4de18,0x800,0x100000)/File(\EFI\debian\grubaa64.efi)
<ardb> results in a filename in the  EFIDP_MEDIA_FILE device node that
       includes the characters that make up File( and )

I checked in the spec, and all it has to say about it is in the examples
of printing a DP section, which says:

 +------------------------------+-----------------------------+
 | Type: 4 (Media Device Path)  | String                      |
 | SubType: 4 (File Path)String | The String is the file path |
 |                              | and is a string.            |
 +------------------------------+-----------------------------+

I've seen little evidence of anything ever using the output of efivar
with the EFI Text to Device Path protocol, so I don't think it's a
problem to make the output here consistent with the spec's big
inconsistency.

This patch removes the "File(" and ")" text from the output.

Signed-off-by: Peter Jones <pjones@redhat.com>