efidp_ipv6_addr: fix gateway size
Unfortunately, the ipv6 address device path type has gateway_ip6_addr
defined as a uint8_t rather than 16 of them.

This patch fixes it to be an array.  While this is theoretically ABI
breaking, it'll only affect things that take a sizeof() on that field or
incorrectly take a sizeof() on the struct, both of which are bad to do
in any case, as anything iterating device paths or taking the size of
the path node should be using the size field in the header.

Signed-off-by: Peter Jones <pjones@redhat.com>