Allow -a to create a non-existent variable
There is no much point in failing when trying to append, we may
as well create the variable. This way it can be used for dbx
update payloads that must be applied in append mode even if there
isn't a dbx yet.

This is nicer than using -w with EFI_VARIABLE_APPEND_WRITE added
to the attributes passed with -A

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>