linux: handle non-ACPI systems in device_get()
/sys/devices/pci0000:00/firmware_node/hid does not exist on all systems.
Kernel parameter acpi=off has been reported as one possible reason in
https://github.com/rhboot/efivar/issues/120
But this is not the only case. On systems using device trees you may also
not find this file.

We have multiple parsers in dev_probes[]. It is sufficient that one of
these parsers succeeds to provide the device information.

If for instance pci_root_parser and pci_parser fails, we may still get
information from nvme_parser.

If a parser fails, only write a debug information and continue.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>