Document legal requirements for contributing to this project
In order to contribute to this project, you need to agree to the DCO.
Before this change, there was no good way for someone to know that they
needed to agree to the DCO before they started contributing. New
contributors would be told that they need to agree to the DCO by the
Developer Certificate of Origin GitHub App [1], but that doesn’t happen
until after a new contributor has submitted their pull request.

Closes #284.

[1]: <https://probot.github.io/apps/dco>

Signed-off-by: Jason Yundt <jason@jasonyundt.email>